requirements.txt
opencv-python
PySimpleGUI
Pillow